pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.1",
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "passlib>=1.7.4",
    "psycopg2>=2.9.10",
    "pydantic>=2.11.5",
    "pyjwt>=2.10.1",
    "python-ulid>=3.0.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]
